Properly pulled out helper function
