+++ packages/storage-plus/src/map.rs
@@ -2,16 +2,14 @@ use serde::de::DeserializeOwned;
 use serde::Serialize;
 use std::marker::PhantomData;
 
+use crate::helpers::query_raw;
 use crate::keys::PrimaryKey;
 #[cfg(feature = "iterator")]
 use crate::keys::{EmptyPrefix, Prefixer};
 use crate::path::Path;
 #[cfg(feature = "iterator")]
 use crate::prefix::{Bound, Prefix};
-use cosmwasm_std::{
-    from_slice, to_vec, Addr, Binary, ContractResult, Empty, QuerierWrapper, QueryRequest,
-    StdError, StdResult, Storage, SystemResult, WasmQuery,
-};
+use cosmwasm_std::{from_slice, Addr, QuerierWrapper, StdError, StdResult, Storage};
 
 #[derive(Debug, Clone)]
 pub struct Map<'a, K, T> {
@@ -102,34 +100,6 @@ where
         } else {
             from_slice(&result).map(Some)
         }
-    }
-}
-
-// TODO: move this to a better helpers location
-pub(crate) fn query_raw(
-    querier: &QuerierWrapper,
-    contract_addr: Addr,
-    key: Binary,
-) -> StdResult<Binary> {
-    let request: QueryRequest<Empty> = WasmQuery::Raw {
-        contract_addr: contract_addr.into(),
-        key,
-    }
-    .into();
-
-    let raw = to_vec(&request).map_err(|serialize_err| {
-        StdError::generic_err(format!("Serializing QueryRequest: {}", serialize_err))
-    })?;
-    match querier.raw_query(&raw) {
-        SystemResult::Err(system_err) => Err(StdError::generic_err(format!(
-            "Querier system error: {}",
-            system_err
-        ))),
-        SystemResult::Ok(ContractResult::Err(contract_err)) => Err(StdError::generic_err(format!(
-            "Querier contract error: {}",
-            contract_err
-        ))),
-        SystemResult::Ok(ContractResult::Ok(value)) => Ok(value),
     }
 }
